backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
scikit-learn==1.3.2
shap==0.43.0
opencv-python==4.8.1.78
requests==2.31.0
cloudinary==1.36.0
prometheus-client==0.19.0
mlflow==2.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
fastapi-cors==0.0.6